% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjoint1.R
\name{adjoint1}
\alias{adjoint1}
\title{Compute the adjoint of a square matrix. The adjoint is the transpose of the cofactor matrix}
\usage{
adjoint1(A)
}
\arguments{
\item{A}{a square matrix}
}
\value{
The value matrix of A
}
\description{
Compute the adjoint of a square matrix. The adjoint is the transpose of the cofactor matrix
}
