% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_AME.R
\name{boot_AME}
\alias{boot_AME}
\title{Function to perform the Monte-Carlo simulations displayed in DDL for the AME using parallel computation.}
\usage{
boot_AME(
  s,
  grid_T,
  prop_T,
  n,
  dimX,
  DGP,
  beta0,
  Option = "quick",
  alpha_c_v = 0,
  nb_clust = 1,
  compute_T = NULL,
  ratio = 10
)
}
\arguments{
\item{s}{indexes the  Monte-Carlo simulations}

\item{grid_T}{is the grid of values of maximal periods T observed in the simulated dataset}

\item{prop_T}{is the vector of proportions of invidividals with respective values of T in grid_T.}

\item{n}{is the number of individuals in the simulated dataset}

\item{dimX}{is the number of covariates}

\item{DGP}{is the DGP, with same numeration as in DDL}

\item{beta0}{is the vector of the true values of the parameter beta}

\item{Option}{the method chosen to perform estimation of the bounds}

\item{alpha_c_v}{the hypothetical vector of clustered values for alpha}

\item{nb_clust}{the number of hypothetical clusters}

\item{compute_T}{the vector of selected periods to compute the AME.}

\item{ratio}{the ratio R in DDL for the nonparametric estimator of the conditional moments of S
If NULL, then as described in Section 5.4 of DDL, AME is computed at the minimum of the support of T.
If specified to ``all",  AME is computed at all available periods but the average over the latter
 is also computed. Default is NULL.}
}
\value{
a list of vectors containing the different outcomes of the estimation (see compute_AME).
}
\description{
Function to perform the Monte-Carlo simulations displayed in DDL for the AME using parallel computation.
}
