% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ATE.R
\name{boot_ATE}
\alias{boot_ATE}
\title{Function to perform the Monte-Carlo simulations displayed in DDL for the ATE using parallel computation.}
\usage{
boot_ATE(
  s,
  selectX = 1,
  grid_T,
  prop_T,
  n,
  dimX,
  DGP,
  beta0,
  Option = "quick",
  alpha_c_v = 0,
  nb_clust = 1,
  compute_T = NULL
)
}
\arguments{
\item{s}{indexes the  Monte-Carlo simulations}

\item{selectX}{the selected covariate with respect to which we compute the ATE}

\item{grid_T}{is the grid of values of maximal periods T observed in the simulated dataset}

\item{prop_T}{is the vector of proportions of invidividals with respective values of T in grid_T.}

\item{n}{is the number of individuals in the simulated dataset}

\item{dimX}{is the number of covariates}

\item{DGP}{is the DGP, with same numerotation as in DDL}

\item{beta0}{is the vector of the true values of the parameter beta}

\item{Option}{the method choosen to perform estimation of the bounds}

\item{alpha_c_v}{the hypothetical vector of clustered values for alpha}

\item{nb_clust}{the number of hypothetical clusters}

\item{compute_T}{the vector of selected periods to compute the AME/ATE.
If NULL, then as described in Section 5.4 of DDL, ATE are computed at the mininum of the support of T.
If specified to ``all",  ATE are computed at all available periods but the average over the latter
 is also computed. Default is NULL.}
}
\value{
a list of vectors containing the different outcomes of the estimation (see compute_ATE).
}
\description{
Function to perform the Monte-Carlo simulations displayed in DDL for the ATE using parallel computation.
}
