% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_MC.R
\name{boot_MC}
\alias{boot_MC}
\title{Function to compute the true bounds for the Delta(x) for each individual value of x
 in the simulated examples of DDL using parallel computing.}
\usage{
boot_MC(i, m_true, lambda_T_plus1, T, c_true)
}
\arguments{
\item{i}{indexes the individual}

\item{m_true}{a vector of true moments m(x) of size T}

\item{lambda_T_plus1}{the value of lambda(x) at T+1}

\item{T}{the number of observed periods for this indvidual}

\item{c_true}{a vector of moments c(x) of size T}
}
\value{
the bounds on Delta(x)
}
\description{
Function to compute the true bounds for the Delta(x) for each individual value of x
 in the simulated examples of DDL using parallel computing.
}
