% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_slow.R
\name{boot_slow}
\alias{boot_slow}
\title{Compute the bounds on q_T(x) for each individual value of x for the sharp method in DDL.
Allows to use parallel computing.}
\usage{
boot_slow(
  i,
  m_hat,
  mat_C_k_T,
  RK,
  n,
  dimX,
  T,
  h,
  PSt_X,
  c0,
  C_mat,
  f,
  lambda_T_plus1,
  X
)
}
\arguments{
\item{i}{indexes the individual}

\item{m_hat}{the vector of estimated moments up to T}

\item{mat_C_k_T}{a matrix containing the combinatorial numbers}

\item{RK}{the integral of the kernel K^2 intervening in the estimator of the asymptotic variance of the estimator of gamma}

\item{n}{the sample size}

\item{dimX}{the number of covariates}

\item{T}{the number of observed periods for this indvidual}

\item{h}{the vector of bandwidths for the nonparametric estimator of gamma}

\item{PSt_X}{the estimated matrix of size n x T of estimated gamma0t(x)}

\item{c0}{the vector containing the first component of the moments c_t(x) for all individuals}

\item{C_mat}{the matrix containing the S-elementary polynomials C_S}

\item{f}{an estimator of the density of X at x.}

\item{lambda_T_plus1}{the vector containing the coefficients T+1 of lambda for all individuals}

\item{X}{matrix of containing the values of the covariates}
}
\value{
bounds on q_T(x) for ith individual
}
\description{
Compute the bounds on q_T(x) for each individual value of x for the sharp method in DDL.
Allows to use parallel computing.
}
