% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_mom.R
\name{bound_mom}
\alias{bound_mom}
\title{Bound on the T0+1-th moment given the vector of first moments m.}
\usage{
bound_mom(m, T0, indic)
}
\arguments{
\item{m}{the vector m until m_{T0}}

\item{T0}{the lenght of the vector of moments m}

\item{indic}{We may have length(m)<T0 if m is at the boundary. In this case, this 3rd
argument (indic) is required, with indic=1 if the next element of m is =
   to the upper bound, indic=-1 if it is = to the lower bound}
}
\value{
return the bounds on the T0+1-th moment
}
\description{
Bound on the T0+1-th moment given the vector of first moments m.
}
