% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_next_moment.R
\name{bound_next_moment}
\alias{bound_next_moment}
\title{Bounds the (Tobsd + 1)-th moment given the first Tobsd moments, m. We
first check that the input moment of vectors is valid, and if necessary
replace it by the vector on the boundary of the set of admissible vectors
of moment (see DDL paper).}
\usage{
bound_next_moment(
  m,
  asymptotic_variance_matrix_m = NULL,
  thresholdDeterminant = 0
)
}
\arguments{
\item{m}{is the estimated vector of moments. NAs at the end is ignored.}

\item{asymptotic_variance_matrix_m}{(default NULL) is the estimated variance matrix for
the vector of moments m. If NULL, the moment vector is assumed to be known
with perfect accuracy (no noise).}

\item{thresholdDeterminant}{(default 0). The threshold used to decide whether the
determinant of the Hankel matrices is null. If the estimated determinant is less
than that threshold, it is considered null (so the moment vector is on the boundary).
Setting the threshold as a small positive value is meant to account for the noise on
the moment vector, e.g. to ensure moment vectors on the boundary are rejected exactly
95\% of the time.}
}
\value{
a list containing:
- lowerBound: the lower bound on the (Tobsd + 1)-th moment
- upperBound: the upper bound on the (Tobsd + 1)-th moment
}
\description{
Bounds the (Tobsd + 1)-th moment given the first Tobsd moments, m. We
first check that the input moment of vectors is valid, and if necessary
replace it by the vector on the boundary of the set of admissible vectors
of moment (see DDL paper).
}
