% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_S_fun.R
\name{C_S_fun}
\alias{C_S_fun}
\title{Computes the elementary polynomials for each row of the input, by induction. This corresponds
to the what the function C_S does to V in the DDL paper.}
\usage{
C_S_fun(V)
}
\arguments{
\item{V}{the matrix of size n x T whose coefficients are
used in the elementary symmetric polynomial. NAs can be used for
unobserved values.}
}
\value{
a matrix of size n x (T + 1) whose value in the (i, j)-th position
is C_(j-1)(X, beta) using the values of V = X' * beta in the i-th row. When
j - 1 is larger than the number of non-NA values of V in the relevant row, the
value NA is returned.
}
\description{
Computes the elementary polynomials for each row of the input, by induction. This corresponds
to the what the function C_S does to V in the DDL paper.
}
