% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChebyshevPoly.R
\name{ChebyshevPoly}
\alias{ChebyshevPoly}
\title{Given nonnegative integer n, compute the Chebyshev polynomial T_n.}
\usage{
ChebyshevPoly(n)
}
\arguments{
\item{n}{indexes the Chebyshev polynomial T_n}
}
\value{
Return the result as a column vector whose mth element is the coefficient of x^(n+1-m).
}
\description{
This code is adapted from ChebyshevPoly.m by David Terr, Raytheon, 5-10-04
}
