% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{choose}
\alias{choose}
\title{Compute C(n,k) with n and k scalars
Allows for negative k or n<k (then returns 0)}
\usage{
choose(n, k)
}
\arguments{
\item{n}{scalar in the combinatorial numers C(n,k)}

\item{k}{scalar in the combinatorial numers C(n,k)}
}
\value{
the value of C(n,k)
}
\description{
Compute C(n,k) with n and k scalars
Allows for negative k or n<k (then returns 0)
}
