% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_Cheb01.R
\name{coeff_Cheb01}
\alias{coeff_Cheb01}
\title{Computes the coefficients of T_k(2u-1), where T_k is the usual Chebyshev
polynomial of the 1st kind}
\usage{
coeff_Cheb01(k)
}
\arguments{
\item{k}{parameter k in T_k(2u-1), where T_k is the usual Chebyshev}
}
\value{
the coefficients of T_k(2u-1)
}
\description{
Computes the coefficients of T_k(2u-1), where T_k is the usual Chebyshev
polynomial of the 1st kind
}
