% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_asymptotic_variance_moment_vector.R
\name{compute_asymptotic_variance_moment_vector}
\alias{compute_asymptotic_variance_moment_vector}
\title{This function computes the asymptotic variance on the estimated moment
for a given individual, that results from the (non-parametric) estimation
of the P(S = j | X)'s.}
\usage{
compute_asymptotic_variance_moment_vector(
  PSt_X,
  Xdensity,
  Zt_mat,
  nbIndls,
  dimX,
  h_local_lin_proba,
  m,
  c0
)
}
\arguments{
\item{PSt_X}{a vector of length Tmax + 1 containing in j-th position the
estimate for P(S = j - 1 | X) for the observation at hand.}

\item{Xdensity}{the estimated density for the covariates (X_1, ..., X_T) for
the observation at hand}

\item{Zt_mat}{a matrix of size (Tmax + 1) x (Tmax + 1) containg in position
(i, j) the value of Z_i(X, j, beta) for the observation at hand. Thus,
c = Zt_mat * PSt_X and m = (c_1, ..., c_T) / c_0, from which we deduce the
asymptotic variance.}

\item{nbIndls}{the total number of individuals used to estimate the
P(S = j | X)'s.}

\item{dimX}{the total number of covariates at a given period.}

\item{h_local_lin_proba}{vector of length Tmax + 1 where the j-th element is
the bandwidth used to estimate P(S = j - 1 | X).}

\item{m}{estimated moment vector (length Tmax) for the observation at hand.
Starts with the 1st moment.}

\item{c0}{estimated value of c_0.}
}
\value{
a matrix of size Tmax x Tmax, an empirical estimate for the value
of the asymptotic variance of the vector of moments m.
}
\description{
This function computes the asymptotic variance on the estimated moment
for a given individual, that results from the (non-parametric) estimation
of the P(S = j | X)'s.
}
