% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_average_AME.R
\name{compute_average_AME}
\alias{compute_average_AME}
\title{Compute the average of the average marginal effects over all the periods between 1 and Tinf.}
\usage{
compute_average_AME(
  output,
  Option,
  Tinf,
  dimX,
  selectX,
  CIOption,
  alpha,
  g_labels,
  G_types,
  G_indic
)
}
\arguments{
\item{output}{a list containing the outputs of the estimation of the bounds on the AME for the periods between 1 and Tinf.}

\item{Option}{chosen method to compute the AME: either sharp or quick. Default is quick.}

\item{Tinf}{vector of size n x1 containing the number of periods observed for each individual.}

\item{dimX}{the number of covariates}

\item{selectX}{the vector of selected covariates to compute the AME.
If NULL then bounds are computed for all covariates. NULL by default.}

\item{CIOption}{the option for the choice of the type of confidence intervals for the quick method, either CI2 or CI3. Default is CI2.}

\item{alpha}{the confidence level for the confidence intervals. Default is 5\%.}

\item{g_labels}{a matrix nx1 containing the individual labels refering to the type of attrition observed and described in the table G_types.}

\item{G_types}{a matrix describing the different possible types of attrition observed in the dataset}

\item{G_indic}{a matrix nxT containing the individual periods observed (0 if unobserved)}
}
\value{
A list containing the values of : Option, n, Tmax,

 - Delta_hat: either the estimator of the bounds on the average of Delta (sharp method) or the approximation of Delta (quick method) over the periods;

 - length_CI: the length of the confidence intervals;

 - et: the estimated standard error of the influence function(s), either of the two bounds on the average of Delta (sharp method) or of
 the approximation of Delta (quick method);

 - bias_sup: the estimated upper bound on the bias (quick method);

 - CI: the confidence intervals at the alpha level;

 - b_hat: the estimated value of beta0 using the conditional maximum likelihood estimator;

 - std_b: the estimated standard deviation of the inflence function of the  conditional maximum likelihood estimator of beta;

 - influence: the matrix of size n x dimX containing the inflence function of the  conditional maximum likelihood estimator of beta;
}
\description{
Compute the average of the average marginal effects over all the periods between 1 and Tinf.
}
