% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_IM_quantile.R
\name{compute_IM_quantile}
\alias{compute_IM_quantile}
\title{Computes the quantile suggested by Imbens & Manski 2004 in order to have
CI with asymptotic level of 1 - alpha. It solves Eq. (2) in Stoye (2009)}
\usage{
compute_IM_quantile(alpha, bounds1, std_bounds1)
}
\arguments{
\item{alpha}{the level for the confidence interval}

\item{bounds1}{a vector of length 2, the estimated bounds}

\item{std_bounds1}{a vector of length 2, the estimated standard errors for
each bound}
}
\value{
the value of the quantile suggested by Imbens & Manski 2004
}
\description{
Computes the quantile suggested by Imbens & Manski 2004 in order to have
CI with asymptotic level of 1 - alpha. It solves Eq. (2) in Stoye (2009)
}
