% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condl_log_lik_FE_logit.R
\name{condl_log_lik_FE_logit}
\alias{condl_log_lik_FE_logit}
\title{Evaluates the conditional log likelihood function of the panel data logit
at beta}
\usage{
condl_log_lik_FE_logit(beta, Y, X)
}
\arguments{
\item{beta}{the value at which the loss is computed}

\item{Y}{a matrix of size n x Tmax containing the values of the dependent
variable Y}

\item{X}{an array of size n x Tmax x dimX containing the values of the
covariates Xs}
}
\value{
the value of the conditional log likelihood function at beta
}
\description{
If Y or X contain NA for some individuals at some periods, these
individual-period pairs are discarded.
}
