% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delta_bounds.R
\name{Delta_bounds}
\alias{Delta_bounds}
\title{This function computes what corresponds to the function under(over)lined h in Section 3.2.2 in DDL}
\usage{
Delta_bounds(beta1, V, S, dimX, PSt_X, f, h, mat_C_k_T, RK, T_minus_t, Tall, X)
}
\arguments{
\item{beta1}{value of the variable beta of the function h (4th parameter in DDL)}

\item{V}{matrix of v = exp(x_t'betahat), where betahat is the estimated value of beta0 using the conditional maximum likelihood estimator.}

\item{S}{matrix of size nx 1 of values of the statistic S}

\item{dimX}{the number of covariates}

\item{PSt_X}{matrix of size n x T containing the estimated conditional probabilities of S given X (gamma in DDL) for all individuals}

\item{f}{vector of size nx 1 containing the estimated values of the density of X evaluated at the observed individual values of X}

\item{h}{vector containing the bandwidths for the nonparametric estimator of the conditional probabilities of S given X}

\item{mat_C_k_T}{matrix containing the combinatorial numbers (T-t, S-t) for all values of t.}

\item{RK}{integral of the kernel squared for the computation of the nonparametric estimator of the conditional probabilities of S given X}

\item{T_minus_t}{matrix of all values of T-t}

\item{Tall}{vector of size n x1 containing the specified values of T where we compute the AME.}

\item{X}{matrix of containing the values of the covariates}
}
\value{
return a vector containing:

    - the average of the functions underline h and overline h over all individuals

    - a matrix of size n times 2 containing the individual values of underline h and overline h
}
\description{
This function computes what corresponds to the function under(over)lined h in Section 3.2.2 in DDL
}
