% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_intercept.R
\name{estim_intercept}
\alias{estim_intercept}
\title{Estimate the intercept of a Logit model of Y on X_T with the
beta estimated by conditional logit}
\usage{
estim_intercept(Y, Xbeta)
}
\arguments{
\item{Y}{a matrix of size nxT to compute the intercept}

\item{Xbeta}{a matrix of size nxT containing bhat'x_t where bhat is the estimated value of beta0 using the conditional maximum likelihood estimator.}
}
\value{
the intercept of a Logit model of Y on X_T with the
beta estimated by conditional logit
}
\description{
Estimate the intercept of a Logit model of Y on X_T with the
beta estimated by conditional logit
}
