% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_intercept_logit.R
\name{estim_intercept_logit}
\alias{estim_intercept_logit}
\title{Estimates the intercept of a logit model on the data}
\usage{
estim_intercept_logit(data)
}
\arguments{
\item{data}{is an environment variable containing the relevant data:
- data$Y a matrix of size n x Tmax containing the values of the
dependent variable Y.
- data$V an array of size n x Tmax containing the values of X'beta
for the estimated value of beta.}
}
\value{
the intercept of a logit model of data$Y on data$X with the
beta estimated by conditional logit
}
\description{
Estimates the intercept of a logit model on the data
}
