% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIndlPeriod.R
\name{extractIndlPeriod}
\alias{extractIndlPeriod}
\title{This function takes as input a matrix M and two vectors, vecIndls and vecPeriods
It returns the vector whose i-th line is given by M[vecIndls[i], vecPeriods[i]].}
\usage{
extractIndlPeriod(M, vecPeriods, vecIndls = NULL)
}
\arguments{
\item{M}{n x T matrix from which we want to extract elements.}

\item{vecPeriods}{vector corresponding to the column numbers of the elements we want to extract.
Can also be a constant, it will then be constant for all elements of vecIndls. NAs are passed on
to the output.}

\item{vecIndls}{vector corresponding to the row numbers of the elements we want to extract. Can
also be a constant, it will then be constant for all elements of vecPeriods. Default is all rows,
once, in order (the vector 1:n).}
}
\value{
If vecPeriods and vecIndls are vectors, returns the vector whose i-th line is given
by M[vecIndls[i], vecPeriods[i]]. If either is a matrix, returns the matrix whose (i, j)-th
element of the output is M[vecIndls[i, j], vecPeriods[i, j]].
}
\description{
Instead of vectors, vecIndls and vecPeriods can also be matrices, in which case
the (i, j)-th element of the output is M[vecIndls[i, j], vecPeriods[i, j]].
}
\details{
If one is a vector and the other a matrix, the vector is interpreted as column
vector standing for a matrix constant on each row (for vecIndls) or as a row vector
standing for a matrix constant for each column (for vecPeriods).
}
