% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gausshermi.R
\name{gausshermi}
\alias{gausshermi}
\title{Returns weights and nodes of the Gauss-Hermite method
Allows for a weight exp(-(x-mu)^2/(2sigma^2)) instead of the classical exp(-x^2)}
\usage{
gausshermi(n, mu, sigma)
}
\arguments{
\item{n}{the maximal degree of the polynomials H_n used in the Gauss-Hermite quadrature method}

\item{mu}{mean centering parameter for the normal weight}

\item{sigma}{standard deviation parameter for the normal weight}
}
\value{
a list containing:

     -  the weights and

     -  the nodes of the Gauss-Hermite method
}
\description{
Returns weights and nodes of the Gauss-Hermite method
Allows for a weight exp(-(x-mu)^2/(2sigma^2)) instead of the classical exp(-x^2)
}
