% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate the simulated data in wide format according to the DGPs in DDL.}
\usage{
generate_data(
  grid_T,
  prop_T,
  n,
  dimX,
  DGP,
  beta0,
  type_cont,
  alpha_c_v = 0,
  nb_clust = 1
)
}
\arguments{
\item{grid_T}{a vector containing the different numbers of periods observed in the dataset.}

\item{prop_T}{a vector containin the proportions of the different numbers of periods observed in the dataset.}

\item{n}{the sample size.}

\item{dimX}{the number of covariates.}

\item{DGP}{the number indexing the chosen DGP, specified as in DDL.}

\item{beta0}{a vector containing the true values of the parameter beta.}

\item{type_cont}{a vector of dummies coding for the different types of the covariates, either continous (1) or binary (0).}

\item{alpha_c_v}{a vector of length the number of specified clusters specifying the average value of alpha in these clusters.}

\item{nb_clust}{the number of hypothetical clusters.}
}
\value{
a list containing:

- Xall: an array of size n x T x dimX containing the simulated values of the different covariates at all periods for all individuals;

- Yall: an matrix of size n x T containing the simulated values of the dependent variable Y at all periods for all individuals;

- Call: a vector of length n containing the index of the hypothetical clusters for all individuals. Default is a vector of ones.
}
\description{
Generate the simulated data in wide format according to the DGPs in DDL.
}
