% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermipol.R
\name{hermipol}
\alias{hermipol}
\title{Function conding the Hermite polynomial of degree n}
\usage{
hermipol(n)
}
\arguments{
\item{n}{degree of the Hermite polynomial.}
}
\value{
the coefficients of the Hermite polynomial of degree n.
}
\description{
Function conding the Hermite polynomial of degree n
}
