% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infl_func_beta.R
\name{infl_func_beta}
\alias{infl_func_beta}
\title{Compute the influence function of beta hat: E(d^2ln L/dbeta^2)^{-1}  x (ln L/dbeta)(Yi|Xi,beta).}
\usage{
infl_func_beta(beta, Y, X, Call = NULL)
}
\arguments{
\item{beta}{the parameter at which the influence function is computed}

\item{Y}{a matrix of size nxT containing the values of the dependent variable Y}

\item{X}{an array of size nxTxdimX containing the values of the covariates X}

\item{Call}{a vector of size n containing the indexes of the hypothetical clusters for the individuals}
}
\value{
a vector of size nx1 containing the value of the influence function
}
\description{
Compute the influence function of beta hat: E(d^2ln L/dbeta^2)^{-1}  x (ln L/dbeta)(Yi|Xi,beta).
}
