% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnot.R
\name{isnot}
\alias{isnot}
\title{Count the non NA values of a vector}
\usage{
isnot(x)
}
\arguments{
\item{x}{a vector, potentially containing NA}
}
\value{
the number of non NA coefficients of x
}
\description{
Count the non NA values of a vector
}
