% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lambda.R
\name{Lambda}
\alias{Lambda}
\title{Evaluate the logit function and its derivative}
\usage{
Lambda(x, deriv = NULL)
}
\arguments{
\item{x}{value at which the logit is evaluated}

\item{deriv}{if not null, evaluate the derivative at x instead of the function}
}
\value{
the value of the logit function at x.
}
\description{
Evaluate the logit function and its derivative
}
