% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lambda_prime.R
\name{Lambda_prime}
\alias{Lambda_prime}
\title{Evaluate the derivative of logit function}
\usage{
Lambda_prime(x)
}
\arguments{
\item{x}{value at which the derivative of logit function is evaluated}
}
\value{
the value of the derivative of logit function at x
}
\description{
Evaluate the derivative of logit function
}
