% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_reg.R
\name{lin_reg}
\alias{lin_reg}
\title{Evaluate the mean square error loss function of linear regression by pooling, to obtain a starting value for the conditional maximum likelihood estimator.}
\usage{
lin_reg(beta, Y, X)
}
\arguments{
\item{beta}{the value at which the loss is computed}

\item{Y}{a matrix of size nxT containing the values of the dependent variable Y}

\item{X}{an array of size nxTxdimX containing the values of the covariates Xs}
}
\value{
the value of the mean square error loss function of linear regression at beta
}
\description{
Evaluate the mean square error loss function of linear regression by pooling, to obtain a starting value for the conditional maximum likelihood estimator.
}
