% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_FE_logit.R
\name{log_lik_FE_logit}
\alias{log_lik_FE_logit}
\title{Evaluate the conditional loglikelihood function of the panel data logit at beta}
\usage{
log_lik_FE_logit(beta, Y, X)
}
\arguments{
\item{beta}{the value at which the loss is computed}

\item{Y}{a matrix of size nxT containing the values of the dependent variable Y}

\item{X}{an array of size nxTxdimX containing the values of the covariates Xs}
}
\value{
the value of the conditional loglikelihood function at beta
}
\description{
If Y or X contain NA for some individuals at periods, these individuals-periods only are discarded.
}
