% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{omega}
\alias{omega}
\title{Evaluate the function Omega in DDL (see Section 4.1), with v the vector of v_t for t
less than T0}
\usage{
omega(u, v, T0)
}
\arguments{
\item{u}{the variable u in Omega, which is a polynomial in u}

\item{v}{v the vector of v_t  (often equals to exp(x_t'beta) or exp((x_t-x_T0)'beta)) for t less than T0}

\item{T0}{T0 +1 is the degree in u of the polynomial Omega}
}
\value{
the value of Omega at u,v
}
\description{
Evaluate the function Omega in DDL (see Section 4.1), with v the vector of v_t for t
less than T0
}
