% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_IM.R
\name{quantile_IM}
\alias{quantile_IM}
\title{Computes the quantile suggested by Imbens & Manski 2004 in order to have
CI with as. size of 1-alpha. It solves Eq. (2) in Stoye (2009)}
\usage{
quantile_IM(alpha, bounds1, std_bounds1)
}
\arguments{
\item{alpha}{the level for the confidence intervals}

\item{bounds1}{the estimated bounds}

\item{std_bounds1}{the estimated standard errors for the bounds}
}
\value{
the value of the quantile suggested by Imbens & Manski 2004
}
\description{
Computes the quantile suggested by Imbens & Manski 2004 in order to have
CI with as. size of 1-alpha. It solves Eq. (2) in Stoye (2009)
}
