% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_felogit.R
\name{summary_felogit}
\alias{summary_felogit}
\title{Produce the final summary table for the output of the felogit function}
\usage{
summary_felogit(output, format = NULL)
}
\arguments{
\item{output}{the output of the felogit function}

\item{format}{can take value "latex" to print the latex table}
}
\value{
a kableExtra or xtable table plotted respectively in the R viewer or terminal
}
\description{
Produce the final summary table for the output of the felogit function
}
