% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumNA.R
\name{sumNA}
\alias{sumNA}
\title{Sum only on non NA values}
\usage{
sumNA(x)
}
\arguments{
\item{x}{a vector, potentially containing NA}
}
\value{
the sum over the non NA coefficients of x
}
\description{
Sum only on non NA values
}
