% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_param_bounds_paral_ATE.R
\name{true_param_bounds_paral_ATE}
\alias{true_param_bounds_paral_ATE}
\title{Approximation of the true effect and bounds on the ATE for the DGP in DDL}
\usage{
true_param_bounds_paral_ATE(
  n_l,
  case_DGP,
  T,
  beta0,
  mat_C_k_T,
  nbCores,
  selectX,
  alpha_c = 0
)
}
\arguments{
\item{n_l}{the sample size used to compute the true bounds by Monte Carlo simulations (recommended 10e6)}

\item{case_DGP}{the number indexing the DGP in DDL}

\item{T}{the number of observed periods}

\item{beta0}{a vector for the true values of the parameter beta}

\item{mat_C_k_T}{a matrix containing the combinatorial numbers}

\item{nbCores}{the number of cores used by the program to compute the true bounds on the AME. We strongly recommend to use nbCores>3.}

\item{selectX}{the selected covariate to evaluate the ATE}

\item{alpha_c}{a value to change the average of alpha. Default is 0.}
}
\value{
A list containing:

 - the value of the true bounds on the AME;

 - the true value of the AME.
}
\description{
Approximation of the true effect and bounds on the ATE for the DGP in DDL
}
