% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_m.R
\name{trunc_m}
\alias{trunc_m}
\title{This function truncates m_hat until it belongs to the moment space. The
truncated moment is put in m_trun.
Sigma is an estimator of the "variance" of m_hat, or more accurately, of
the asymptotic variance of m_hat, properly normalized.}
\usage{
trunc_m(m_hat, Sigma = NULL, n = NULL, ind_Sig_pos = NULL)
}
\arguments{
\item{m_hat}{the estimated sequence of moments}

\item{Sigma}{is an estimator of the "variance" of m_hat, or more accurately, of the asymptotic variance of m_hat, properly normalized.}

\item{n}{the sample size}

\item{ind_Sig_pos}{indicates if Sigma is invertible

 @return A list containing:

   - m_trun: the truncated value of m_hat, which belongs to the moment space;

   - indic=0 if m_hat was not truncated. Otherwise, indic=1 if we are above the upper bound, indic=-1 if we are  below the lower bound.}
}
\description{
This function truncates m_hat until it belongs to the moment space. The
truncated moment is put in m_trun.
Sigma is an estimator of the "variance" of m_hat, or more accurately, of
the asymptotic variance of m_hat, properly normalized.
}
