% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_approx_poly.R
\name{best_approx_poly}
\alias{best_approx_poly}
\title{This function returns the coefficients (functions of v=exp(x'beta0)) of
the best (in the uniform sense) approximation  of Omega(.,x) by a
polynomial of degree T.}
\usage{
best_approx_poly(Vtilde_min1, grid_T1, Tinf0)
}
\arguments{
\item{Vtilde_min1}{is the vector of coefficients exp((x-x_T)'beta0)-1 appearing
in Omega}

\item{grid_T1}{a vector containing the different values of the maximal number of periods observed}

\item{Tinf0}{a vector of size nx1 containing the number of periods observed for each individual}
}
\value{
the coefficients of the best approximation of Omega(.,x) by polynomial of degree T.
}
\description{
This function returns the coefficients (functions of v=exp(x'beta0)) of
the best (in the uniform sense) approximation  of Omega(.,x) by a
polynomial of degree T.
}
