% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_S_fun.R
\name{C_S_fun}
\alias{C_S_fun}
\title{Computes by induction one of the elementary symm. polynomial on each row
of V1, matrix of size n x T.}
\usage{
C_S_fun(S1, V1)
}
\arguments{
\item{S1}{the vector of values for the statistic S.}

\item{V1}{the matrix of size n x T which coefficients are
used in the elementary symm. polynomial.}
}
\value{
a vector if S1 is scalar, or a matrix otherwise, containing
the elementary symm. polynomial on each row of V1.
}
\description{
If S1 is scalar, computes the S1-th elem. symm. polynomial
If S1 is a n-column vector, computes, for row i, the S1(i)-th elem. symm.
polynomial.
}
