% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{choose}
\alias{choose}
\title{Takes input n, k and returns n choose k}
\usage{
choose(n, k)
}
\arguments{
\item{n}{number of elements to choose from}

\item{k}{number of elements to choose}
}
\value{
n choose k
}
\description{
Takes input n, k and returns n choose k
}
