% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_AME_t.R
\name{compute_AME_t}
\alias{compute_AME_t}
\title{Function which compute the AME at a specified value of T contained in Tall.}
\usage{
compute_AME_t(
  Yall,
  Xall,
  prop_T,
  grid_T,
  n_dist,
  Tmax,
  Tall,
  Tinf,
  Call1,
  mat_C_k_T,
  cheb_coeff,
  b_hat,
  alpha = 0.05,
  CIOption,
  Option,
  dimX,
  selectX = NULL,
  phi_b,
  std_b,
  nbCores = 4,
  ratio = 10
)
}
\arguments{
\item{Yall}{matrix n x T containing the values of the dependent variable Yt.}

\item{Xall}{array of dimensions n x T x dimX containing the values of the predictors at the different periods.}

\item{prop_T}{a vector containin the proportions of the different numbers of periods observed in the dataset.}

\item{grid_T}{a vector containing the different numbers of periods observed in the dataset.}

\item{n_dist}{a vector containing the number of individuals which are observed for T periods, where T is given in grid_T}

\item{Tmax}{the maximal number of periods observed in the dataset}

\item{Tall}{vector of size n x1 containing the specified values of T where we compute the AME.}

\item{Tinf}{vector of size n x1 containing the number of periods observed for each individual.}

\item{Call1}{matrix n x 1 containing the identifiers of the clusters to which each individual belongs. Default is NULL}

\item{mat_C_k_T}{a matrix containing the combinatorial numbers.}

\item{cheb_coeff}{the coefficients of the Chebyshev polynomials T_n.}

\item{b_hat}{the estimated value of beta0 using the conditional maximum likelihood estimator.}

\item{alpha}{the confidence level for the confidence intervals. Default is 5\%.}

\item{CIOption}{the option for the choice of the type of confidence intervals for the quick method, either CI2 or CI3. Default is CI2.}

\item{Option}{chosen method to compute the AME: either sharp or quick. Default is quick.}

\item{dimX}{the number of covariates}

\item{selectX}{the vector of selected covariates to compute the AME.
If NULL then bounds are computed for all covariates. NULL by default.}

\item{phi_b}{the matrix of size n x dimX containing the influence function of the  conditional maximum likelihood estimator of beta.}

\item{std_b}{the estimated standard deviation of the influence function of the  conditional maximum likelihood estimator of beta.}

\item{nbCores}{the number of cores used by the program to compute the AME for the ``sharp" method.}

\item{ratio}{the ratio R in DDL for the nonparametric estimator of the conditional moments of S}
}
\value{
A list containing the values of : Option, n, Tmax, Time (computational time),

 - Delta_hat: either the estimator of the bounds on Delta (sharp method) or the approximation of Delta (quick method);

 - length_CI: the length of the confidence intervals;

 - et: the estimated standard error of the influence function(s), either of the two bounds on Delta (sharp method) or of
 the approximation of Delta (quick method);

 - bias_sup: the estimated upper bound on the bias (quick method);

 - CI: the confidence intervals at the alpha level;

 - phi_alpha: the phi alpha statistic intervening in the computation of the confidence interval of the sharp method;

 - b_hat: the estimated value of beta0 using the conditional maximum likelihood estimator;

 - std_b: the estimated standard deviation of the influence function of the  conditional maximum likelihood estimator of beta;

 - influence: the matrix of size n x dimX containing the influence function of the  conditional maximum likelihood estimator of beta;

 - for_average: a list containing the estimated values of Delta(x) and the influence function weighted by the values of Tinf, to compute the
 average of the AME over all the observed periods.
}
\description{
Function which compute the AME at a specified value of T contained in Tall.
}
