% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ATE.R
\name{compute_ATE}
\alias{compute_ATE}
\title{Function which compute the ATE at different values of T according to the selected values of compute_T}
\usage{
compute_ATE(
  Yall,
  Xall,
  Call = NULL,
  Option = "quick",
  selectX = 1,
  compute_T = NULL,
  alpha = 0.05,
  CIOption = "CI2",
  g_labels = NULL,
  G_types = NULL,
  G_indic = NULL
)
}
\arguments{
\item{Yall}{matrix n x T containing the values of the dependent variable Yt}

\item{Xall}{array of dimensions n x T x dimX containing the values of the predictors at the different periods.}

\item{Call}{matrix n x 1 containing the identifiers of the clusters to which each individual belongs. Default is NULL}

\item{Option}{chosen method to compute the ATE: either sharp or quick. Default is quick.}

\item{selectX}{the vector of selected covariates to compute the AME.
If NULL then bounds are computed for all covariates. NULL by default.}

\item{compute_T}{the vector of selected periods to compute the ATE.
If NULL, then as described in Section 5.4 of DDL, ATE is computed at min supp (T).
If specified to ``all",  the ATE is computed at all available periods but the average over the latter is also computed. Default is NULL.}

\item{alpha}{the confidence level for the confidence intervals. Default is 5\%.}

\item{CIOption}{the option for the choice of the type of confidence intervals for the quick method, either CI2 or CI3. Default is CI2.}

\item{g_labels}{a matrix nx1 containing the individual labels refering to the type of attrition observed and described in the table G_types.}

\item{G_types}{a matrix describing the different possible types of attrition observed in the dataset}

\item{G_indic}{a matrix nxT containing the individual periods observed (0 if unobserved)}
}
\value{
a list of all the outputs of compute_ATE_t containing the different results of the estimation of the different values of T considered.
}
\description{
Function which compute the ATE at different values of T according to the selected values of compute_T
}
