% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felogit.R
\name{felogit}
\alias{felogit}
\title{This function implements the estimators of the bounds on the AME/ATE proposed in DDL.}
\usage{
felogit(
  formul = NULL,
  data,
  Option = "quick",
  compute_X = NULL,
  compute_T = NULL,
  cluster = NULL,
  alpha = 0.05,
  CIOption = "CI2",
  nbCores = 4,
  ratio = 10
)
}
\arguments{
\item{formul}{an object of class "formula": a symbolic description of the model to be fitted.
Models for felogit are specified symbolically. A typical model has the form "response ~ terms`" where response is the (numeric) response vector
and terms is a series of terms which specifies a linear predictor for response. Note that predictors which are constant over time are discarded.}

\item{data}{a data frame containing the panel data in long, i.e., one line for one individual-time observation.
The first column must be the individual identifier and the second column the temporal one. It can contains NA, which are treated as explained in the vignette.}

\item{Option}{the method chosen to compute the AME/ATE, either "sharp" or "quick". Default is ``quick".}

\item{compute_X}{the vector of selected covariates to compute the AME/ATE. If NULL then bounds are computed for all covariates. NULL by default.}

\item{compute_T}{the vector of selected periods to compute the AME/ATE.
If NULL, then as described in Section 5.4 of DDL, AME/ATE are computed at the minimum of  supp(T). If specified to ``all",
AME/ATE are computed at all available periods but the average over the latter is also computed. Default is NULL.}

\item{cluster}{the vector of identifiers for the clusters if any.  Default is NULL.}

\item{alpha}{the confidence level for the confidence intervals. Default is 5\%.}

\item{CIOption}{the option for the choice of the type of confidence intervals for the "quick" method, either "CI2" or "CI3". Default is ``CI2".}

\item{nbCores}{the number of cores used by the program to compute the AME for the "sharp" method.}

\item{ratio}{the ratio R in DDL for the nonparametric estimator of the conditional moments of S
To reduce the computational time, this function can use several cores, in which case the library snowfall should be loaded first. By default, nbCores is set to 4.}
}
\value{
A list containing:

 - summary: a dataframe containing the estimation results,

 - n: the number of used individuals,

 - ndiscard: the number of discarded individuals,

 - Tmax: the maximal number of periods of observed,

 - vardiscard: the label of the discarded variables,

 - formul: the formula used,

 - alpha: the level used for the confidence intervals,

 - Option: the method which is used.

 - summary_CMLE : a dataframe containing the estimation results of the CMLE
}
\description{
The function summary_felogit applied to the output of this felogit function prints the table containing the estimation results.
}
\examples{
library(pglm)
data('UnionWage', package = 'pglm')

UnionWage$union = (UnionWage$union =="yes")*1
UnionWage$rural = (UnionWage$rural =="yes")*1
UnionWage$black = (UnionWage$com =="black")*1 # used as test for discarded variable because constant
UnionWage$NorthEast =(UnionWage$region =="NorthEast")*1
sub <- UnionWage[UnionWage$year <1986,]
formul = as.formula(union ~  exper + married +black )
output <- felogit(formul , data= sub, Option  ="quick", compute_T =  NULL )
summary_felogit(output)

}
