% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_lin.R
\name{local_lin}
\alias{local_lin}
\title{Estimates E(Y|X) at the points Xb by local linear regression
with bandwidth h1 and the Epanechnikov kernel.
Second output f: Kernel estimator of the density}
\usage{
local_lin(Yb, Xb, h1)
}
\arguments{
\item{Yb}{a matrix of size nxT containing the values of the dependent variable Y.}

\item{Xb}{an array of size nxTxdimX containing the values of the covariates Xs.}

\item{h1}{bandwidth for the local linear regression.}
}
\value{
a list containing

 - res: the estimates E(Y|X) at the points Xb by local linear regression;

 - f: Kernel estimator of the density.
}
\description{
Estimates E(Y|X) at the points Xb by local linear regression
with bandwidth h1 and the Epanechnikov kernel.
Second output f: Kernel estimator of the density
}
